requirements.txt
numpy==1.22.4
openpyxl==3.0.9
pandas==1.3.3
Pillow==11.0.0
streamlit==1.23.1
streamlit_authenticator==0.1.5
streamlit_option_menu==0.3.2
